Move Ignore annotation into annotations package
