+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/Identifiers.kt
@@ -82,7 +82,7 @@ internal object FqNames {
     val COLLECTION_TYPE = FqName("io.realm.interop.CollectionType")
     val PROPERTY_FLAG = FqName("io.realm.interop.PropertyFlag")
     val PRIMARY_KEY_ANNOTATION = FqName("io.realm.PrimaryKey")
-    val IGNORE_ANNOTATION = FqName("io.realm.Ignore")
+    val IGNORE_ANNOTATION = FqName("io.realm.annotations.Ignore")
     val TRANSIENT_ANNOTATION = FqName("kotlin.jvm.Transient")
     // Realm data types
     val REALM_LIST = FqName("io.realm.RealmList")
